codebase reorganization
