+++ package.json
@@ -10,8 +10,8 @@
     "express",
     "charts"
   ],
-  "engines" : {
-    "node" : ">=4"
+  "engines": {
+    "node": ">=4"
   },
   "author": "Rafal Wilinski <raf.wilinski@gmail.com> (http://rwilinski.me)",
   "contributors": [
